move files into dirs and split
