+++ src/main.ts
@@ -1,2 +1,3 @@
-export { rsc, devServer } from "./lib/middleware.js";
+export { rsc } from "./lib/middleware/rsc.js";
+export { devServer } from "./lib/middleware/devServer.js";
 export { build } from "./lib/builder.js";
